Moved icon library file to top level
