+++ goco-transit/src/components/add-offer-dialog.js
@@ -18,7 +18,7 @@ import Avatar from 'material-ui/Avatar';
 import TextField from 'material-ui/TextField';
 
 // Components
-import {Icons} from './icon-library';
+import {Icons} from '../icon-library';
 
 {/* Add an offer dialog box */}
