+++ goco-transit/src/pages/offers-page.js
@@ -14,7 +14,7 @@ import Badge from 'material-ui/Badge';
 // Components
 import AddOfferDialog from '../components/add-offer-dialog';
 import DeleteOfferDialog from '../components/delete-offer-dialog';
-import {Icons} from '../components/icon-library';
+import {Icons} from '../icon-library';
 
 // Services
 import { getUser } from '../services/user-service';
