+++ goco-transit/src/pages/request-search-page.js
@@ -24,7 +24,7 @@ import Dialog, {
 
 // Components
 import AddRequestDialog from '../components/add-request-dialog';
-import {Icons} from '../components/icon-library';
+import {Icons} from '../icon-library';
 
 // Services
 import { findOfferedRides } from '../services/ride-service';
