+++ goco-transit/src/pages/settings-page.js
@@ -6,7 +6,7 @@ import Checkbox from 'material-ui/Checkbox';
 import { Redirect } from 'react-router'
 
 // Components
-import {Icons} from '../components/icon-library';
+import {Icons} from '../icon-library';
 
 // Services
 import { signOut, isAuthenticated } from '../services/auth-service';
